requirements.txt
python-twitter
scipy
numpy
sklearn
nltk
bokeh